Fix unclear Japanese translations (#2127)
Fix not clear Japanese translation 'Pending Vote' in basetriggers.phrases.txt
Changed `ペンディング投票` to `投票待ち`. Reason is below.
`ペンディング` is the Japanese pronunciation of `Pending` as it is. And not a common phrase in Japanese.
And literal translation of `Pending` is `保留中`. But `保留中` is means `Waiting, but currently not planned.` in most of situation.
But If I understand correctly this context is waiting for a next vote until time/round limit reached or RTV.
So I've changed `ペンディング` to `待つ`. `待つ` means `Wait` in English. And then result of the change to conform to Japanese grammar is `投票待ち`.

Fix not clear Japanese translation 'Slot reserved' in reservedslots.phrases.txt
Changed `予約スロットのために落とされます。` to `予約スロットにより切断されました。`. Reason is below.
`落とす` in `落とされます` is means same as `Drop` in English.
`落とす` have a lot of meanings in one word in Japanese. For example: `Drop` `Disconnect` `To download (software)` `To reduce (speed)` `To lose (purse)` `To shed (leaves)`
So `落とす` is not clear in this context so I've changed `落とす` to `切断`.
Also `されます` is used for things that are scheduled or about to take place. And `されました` is used for something that has already been done.
This message is shown in after disconnect, so I've changed `されます` to `されました`.
Finally, result of the change become `予約スロットにより切断されました。`.